docs: clarify webhook url segments and include port
When deploying this webhook service, we noticed that the url specified
in the docs was missing the port which is assigned by default in the
code. If left unspecified, http will assume port 80, however the code
defaults to port 8000 unless explicitly set otherwise. This also
clarifies the strucuture of the url segments in the webhook.